:recycle: moves getAccessControlFilter to own file
